Have the manager keep the state of host
When sending output reports, the host should maintain the state,
otherwise, the event can reset other event's state.

For example, when an off-hook is triggered first, the following mute
or ring event should reset the off-hook state but to keep it as true.

Also add the State column to the test app to show host's states.